+++ components/basecomponents/MemoryGame/MemoryGame.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useMemo } from 'react';
+import React, { useState, useEffect, useMemo, useCallback } from 'react';
 import { Button } from 'components/basecomponents/Button';
 import { useTranslation } from 'next-i18next';
 import Card from './MemoryGameCard';
@@ -22,7 +22,6 @@ const addBackroundColor = (cardsData: CardData[]) =>
 
 const createCardDeck = (cardsData: CardData[]) => {
   // prepare and shuffle cards, pick 8 cards
-  console.log('creating carddeck');
   const pickedCards = cardsData.sort(() => Math.random() - 0.5).slice(0, 8);
   const coloredCards = addBackroundColor(pickedCards) as (CardData & { color: string })[];
 
@@ -66,17 +65,6 @@ enum Scene {
   goNewGame = 'goNewGame',
 }
 
-enum Transition {
-  begin = 'begin',
-  game = 'game',
-  firstCardSelected = 'firstCardSelected',
-  secondCardSelected = 'secondCardSelected',
-  cardsMatch = 'cardsMatch',
-  cardsDontMatch = 'cardsDontMatch',
-  win = 'win',
-  restart = 'restart',
-}
-
 export type CardData = {
   image: string;
   translation: TranslationJSON;
@@ -101,292 +89,156 @@ export type Theme = {
   cardsData: CardData[];
 };
 
-interface SelectedCards {
-  first: Card | null;
-  second: Card | null;
-}
-
 interface MemoryGameProps {
   theme: Theme;
 }
 
-const createUseGame = () => {
-  interface State {
-    theme: Theme;
-    cards: Card[];
-    selectedCards: SelectedCards;
-    scene: Scene;
-    controlsEnabled: boolean;
+const MemoryGame = ({ theme }: MemoryGameProps) => {
+  const { playCardPhrase, playPhraseRandomLang } = usePlayPhrase();
+  const { t } = useTranslation();
+  const [cards, setCards] = useState<Card[]>([]);
+  const { audio, image, styles, cardsData } = theme;
+
+  interface SelectedCards {
+    first: Card | null;
+    second: Card | null;
   }
 
-  let state: State | null = null;
-  
-  let playCardPhrase: ((card: Card) => Promise<void>) = () => Promise.resolve();
-  let playPhraseRandomLang: ((phrase: TranslationJSON) => Promise<void>) = () => Promise.resolve();
+  const [selectedCards, setSelectedCards] = useState<SelectedCards>({ first: null, second: null });
 
-  const setPlayPhrase = (playPhrase:{playCardPhrase: (card: Card) => Promise<void>,playPhraseRandomLang: (phrase: TranslationJSON) => Promise<void>}) => {
-    playCardPhrase = playPhrase.playCardPhrase;
-    playPhraseRandomLang = playPhrase.playPhraseRandomLang;
-  }
+  const isSelected = (card: Card) =>
+    (selectedCards.first !== null && card.id === selectedCards.first.id) ||
+    (selectedCards.second !== null && card.id === selectedCards.second.id);
 
-  const initializeState = (initState: State) => {
-    console.log('initialize state');
-    state = initState;
-    console.log(state);
-  };
-
-  const setState = (func: (state: State) => State) => {
-    console.log('update state');
-    if (state === null) throw new Error('state not initialized');
-    state = { ...state, ...func(state) };
-    console.log(state);
-  };
-
-  const getState = () => {
-    if (state === null) throw new Error('state not initialized');
-    return state;
-  };
-
-  const [setTimer, clearTimers] = createTimer();
-
-  const isSelected = (card: Card) => {
-    const { selectedCards } = getState();
-    return (
-      (selectedCards.first !== null && card.id === selectedCards.first.id) ||
-      (selectedCards.second !== null && card.id === selectedCards.second.id)
-    );
-  };
-
-  const setCards = (cards: Card[]) => {
-    setState((state) => ({ ...state, cards }));
-  };
-
-  const setScene = (scene: Scene) => {
-    setState((state) => ({ ...state, scene }));
-  };
-
-  const setSelectedCards = (selectedCards: SelectedCards) => {
-    setState((state) => ({ ...state, selectedCards }));
-  };
-
-  const disableControls = () => {
-    setState((state) => ({ ...state, controlsEnabled: false }));
-  };
-
-  const enableControls = () => {
-    setState((state) => ({ ...state, controlsEnabled: true }));
-  };
+  const [scene, setScene] = useState<Scene>(Scene.init);
+  const [controlsDisabled, setControlsDisabled] = useState<boolean>(true);
+  const [resetDisabled, setResetDisabled] = useState<boolean>(false);
+  const [setTimer, clearTimers] = useMemo(createTimer, []);
 
   const flipCard = (cardToFlip: Card) => {
-    console.log('flipp card');
-    setCards(getState().cards.map((card) => (card.id === cardToFlip.id ? { ...card, flipped: !card.flipped } : card)));
+    setCards((cards) => cards.map((card) => (card.id === cardToFlip.id ? { ...card, flipped: !card.flipped } : card)));
   };
 
-  const initGame = (theme: Theme) => {
-    console.log('init game');
-    // init/replace game state with new data
-    const cards = createCardDeck(theme.cardsData);
-
-    const initState = {
-      theme,
-      cards,
-      selectedCards: { first: null, second: null },
-      scene: Scene.init,
-      controlsEnabled: false,
-    };
-
-    initializeState(initState);
+  const game = () => {
+    setScene(Scene.game);
+    setControlsDisabled(false);    
   };
 
-  const handleNewGameClick = () => {
-    console.log('new game click');
-    transitions[Transition.restart]();
+  const win = async () => {    
+    setScene(Scene.win);
+    await playPhraseRandomLang(getRandomElement(phrases.win));
+    await delay(200);
+    setScene(Scene.winReward);
+    playAudio(audio.winMusic);
   };
 
+  const firstCardSelected = async (card: Card) => {
+    setControlsDisabled(true);
+    setSelectedCards({ first: card, second: null });
+    setScene(Scene.firstCardSelected);
+    // play css animations and sounds
+    flipCard(card); // 0.3s
+    await playAudio(audio.cardFlipSound);
+    playCardPhrase(card);
+
+    setTimer(() => {
+      game();
+    }, 600); // reduced for fastrer UX
+  };
+
+  const cardsMatch = async () => {
+    setScene(Scene.cardsMatch);
+    delay(100);
+    await playAudio(audio.cardsMatchSound);
+    setScene(Scene.cardsMatchReward);
+    Math.random() > 0.5 && (await playPhraseRandomLang(getRandomElement(phrases.good)));
+    // reset selected cards
+    // check win
+    if (cards.every((card) => card.flipped)) {
+      win();
+    } else {
+      setSelectedCards({ first: null, second: null });
+      game();
+    }
+  };
+
+  const cardsDontMatch = async ({ first, second }: { first: Card; second: Card }) => {
+    setResetDisabled(true);
+    setScene(Scene.cardsDontMatch);
+    await delay(1000);
+    Math.random() > 0.8 && (await playPhraseRandomLang(getRandomElement(phrases.wrong)));
+    flipCard(first);
+    flipCard(second);
+    await playAudio(audio.cardFlipSound);
+    setSelectedCards({ first: null, second: null });
+    setResetDisabled(false);
+    game();
+  };
+  const secondCardSelected = async ({ first, second }: { first: Card; second: Card }) => {
+    setControlsDisabled(true);    
+    setSelectedCards({ first, second });
+    setScene(Scene.secondCardSelected);
+    // play css animations and sounds
+    flipCard(second); // 0.3s
+    await playAudio(audio.cardFlipSound);
+    await playCardPhrase(second);
+
+    // check if cards match
+    if (first.image === second.image) {
+      cardsMatch();
+    } else {
+      cardsDontMatch({ first, second });
+    }
+  };
   const selectCard = (card: Card) => {
-    const state = getState();
-    if (!state.controlsEnabled || isSelected(card) || card.flipped) return;
+    if (controlsDisabled || isSelected(card) || card.flipped) return;
 
-    const { first, second } = state.selectedCards;
+    const { first, second } = selectedCards;
     // select first card
     if (first === null && !card.flipped) {
-      transitions[Transition.firstCardSelected](card);
+      firstCardSelected(card);
       // select second card
     } else if (first !== null && second === null && !card.flipped) {
-      transitions[Transition.secondCardSelected]({ first, second: card });
+      secondCardSelected({ first, second: card });
     }
   };
 
-  const transitions = {
-    begin: () => {
-      setScene(Scene.begin);
-      disableControls();
-      updateUI();
-      setTimer(() => {
-        transitions[Transition.game]();
-      }, 1000);
-    },
-    game: () => {
+  const begin = useCallback(() => {
+    setScene(Scene.begin);
+    setControlsDisabled(true);
+    setCards(createCardDeck(cardsData));
+    setSelectedCards({ first: null, second: null });
+    setTimer(() => {
       setScene(Scene.game);
-      enableControls();
-      updateUI();
-    },
-    firstCardSelected: async (card: Card) => {
-      disableControls();
-      setSelectedCards({ first: card, second: null });
-      setScene(Scene.firstCardSelected);      
-      // play css animations and sounds
-      flipCard(card); // 0.3s
-      updateUI();
-      await playAudio(getState().theme.audio.cardFlipSound);
-      playCardPhrase(card);
+      setControlsDisabled(false);
+    }, 1000);
+  }, [setTimer, cardsData]);
 
-      setTimer(() => {
-        transitions[Transition.game]();
-      }, 600); // reduced for fastrer UX
-    },
-    secondCardSelected: async ({ first, second }: { first: Card; second: Card }) => {
-      disableControls();
-      setSelectedCards({ first, second });
-      setScene(Scene.secondCardSelected);      
-      // play css animations and sounds
-      flipCard(second); // 0.3s
-      updateUI();
-      await playAudio(getState().theme.audio.cardFlipSound);
-      await playCardPhrase(second);
+  const restart = async () => {
+    if (resetDisabled) return; // maybe play oink sound
+    setControlsDisabled(true);
+    playPhraseRandomLang(getRandomElement(phrases.newGame));
+    setScene(Scene.goNewGame);
+    clearTimers();
+    setTimer(begin, 500);
+  };
 
-      const cardsMatch = first.image === second.image;
-
-      if (cardsMatch) {
-        transitions[Transition.cardsMatch]();
-      } else {
-        transitions[Transition.cardsDontMatch]({ first, second });
-      }
-    },
-    cardsMatch: async () => {
-      setScene(Scene.cardsMatch);
-      updateUI();
-      delay(100);
-      await playAudio(getState().theme.audio.cardsMatchSound);
-      setScene(Scene.cardsMatchReward);
-      updateUI();
-      Math.random() > 0.5 && (await playPhraseRandomLang(getRandomElement(phrases.good)));
-      // reset selected cards
-      // check win
-      if (getState().cards.every((card) => card.flipped)) {
-        transitions[Transition.win]();
-      } else {
-        setSelectedCards({ first: null, second: null });
-        transitions[Transition.game]();
-      }
-    },
-    cardsDontMatch: async ({ first, second }: { first: Card; second: Card }) => {
-      setScene(Scene.cardsDontMatch);
-      updateUI();
-      await delay(1000);
-      Math.random() > 0.8 && (await playPhraseRandomLang(getRandomElement(phrases.wrong)));
-      flipCard(first);
-      flipCard(second);
-      updateUI();
-      await playAudio(getState().theme.audio.cardFlipSound);
-      setSelectedCards({ first: null, second: null });
-      transitions[Transition.game]();
-    },
-    win: async () => {
-      setScene(Scene.win);
-      updateUI();
-      await playPhraseRandomLang(getRandomElement(phrases.win));
-      await delay(200);
-      setScene(Scene.winReward);
-      updateUI();
-      playAudio(getState().theme.audio.winMusic);
-    },
-    restart: async () => {
-      disableControls();
-      playPhraseRandomLang(getRandomElement(phrases.newGame));
-      setScene(Scene.goNewGame);
-      updateUI();
+  useEffect(() => {
+    return () => {
       clearTimers();
-      setTimer(() => {
-        initGame(getState().theme);
-        transitions[Transition.begin]();
-      }, 500);
-    },
-  };
+    };
+  }, [clearTimers]);
 
-  interface Controls {
-    selectCard: (card: Card) => void;
-    isSelected: (card: Card) => boolean;
-    handleNewGameClick: () => void;
-  }
-
-  const controls = { selectCard, isSelected, handleNewGameClick };
-
-  interface Game {
-    state: State;
-    controls: Controls;
-  }
-
-  const getGame = () => (state === null ? null : { state, controls });
-
-  let uiCallback: null | (() => void) = null;
-  const connectUI = (func: () => void) => {
-    console.log('connect ui');
-    uiCallback = func;
-  };
-  const disconnectUI = () => (uiCallback = null);
-  const updateUI = () => {
-    if (uiCallback !== null) {
-      console.log('update ui');
-      uiCallback();
-    }
-  };
-
-  const useGame = (theme: Theme) => {
-    const [game, setGame] = useState<Game | null>(null);
-    const { playCardPhrase, playPhraseRandomLang } = usePlayPhrase();
-
-    useEffect(() => {
-      // connect ui -> setGame()
-      setPlayPhrase({ playCardPhrase, playPhraseRandomLang });
-      connectUI(() => setGame(getGame()));
-      return () => {
-        // disconnect ui
-        disconnectUI();
-      };
-    }, []);
-
-    useEffect(() => {
-      // init with new data
-      initGame(theme);
-      transitions[Transition.begin]();
-      // start new game maybe
-    }, [theme]);
-
-    return game;
-  };
-
-  return useGame;
-};
-
-const useGame = createUseGame();
-
-const MemoryGame = ({ theme }: MemoryGameProps) => {  
-  const { t } = useTranslation();
-  
-  const game = useGame(theme);
-  if (game === null) return null;
-
-  const { scene, cards } = game.state;
-  const { image, styles } = game.state.theme;
-  const { isSelected, selectCard, handleNewGameClick } = game.controls;
+  useEffect(() => {
+    clearTimers();
+    begin();
+  }, [theme, begin, clearTimers]);
 
   return (
     <div className={styles.app}>
-      <Button className={styles.newGameButton} text={t('utils.new_game')} onClick={handleNewGameClick} />
+      <Button className={styles.newGameButton} text={t('utils.new_game')} onClick={restart} />
       <div className={styles.board}>
         {scene !== Scene.init &&
-          cards !== null &&
           cards.map((card) => (
             <Card
               key={card.id}
