+++ components/basecomponents/MemoryGame/MemoryGameGlobal.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useMemo } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Button } from 'components/basecomponents/Button';
 import { useTranslation } from 'next-i18next';
 import Card from './MemoryGameCard';
@@ -101,39 +101,116 @@ export type Theme = {
   cardsData: CardData[];
 };
 
+interface SelectedCards {
+  first: Card | null;
+  second: Card | null;
+}
+
 interface MemoryGameProps {
   theme: Theme;
 }
 
-const MemoryGame = ({ theme }: MemoryGameProps) => {
-  const { playCardPhrase, playPhraseRandomLang } = usePlayPhrase();
-  const { t } = useTranslation();
-  const [cards, setCards] = useState<Card[]>([]);
-  const { audio, image, styles, cardsData } = theme;
-  
-  interface SelectedCards {
-    first: Card | null;
-    second: Card | null;
+const createUseGame = () => {
+  interface State {
+    theme: Theme;
+    cards: Card[];
+    selectedCards: SelectedCards;
+    scene: Scene;
+    controlsEnabled: boolean;
   }
 
-  const [selectedCards, setSelectedCards] = useState<SelectedCards>({ first: null, second: null });
+  let state: State | null = null;
 
-  const isSelected = (card: Card) =>
-    (selectedCards.first !== null && card.id === selectedCards.first.id) ||
-    (selectedCards.second !== null && card.id === selectedCards.second.id);
+  let playCardPhrase: (card: Card) => Promise<void> = () => Promise.resolve();
+  let playPhraseRandomLang: (phrase: TranslationJSON) => Promise<void> = () => Promise.resolve();
 
-  const [scene, setScene] = useState<Scene>(Scene.init);
-  const [controlsDisabled, setControlsDisabled] = useState<boolean>(true);
-  const [setTimer, clearTimers] = useMemo(createTimer, []);
+  const setPlayPhrase = (playPhrase: {
+    playCardPhrase: (card: Card) => Promise<void>;
+    playPhraseRandomLang: (phrase: TranslationJSON) => Promise<void>;
+  }) => {
+    playCardPhrase = playPhrase.playCardPhrase;
+    playPhraseRandomLang = playPhrase.playPhraseRandomLang;
+  };
+
+  const initializeState = (initState: State) => {
+    console.log('initialize state');
+    state = initState;
+    console.log(state);
+  };
+
+  const setState = (func: (state: State) => State) => {
+    console.log('update state');
+    if (state === null) throw new Error('state not initialized');
+    state = { ...state, ...func(state) };
+    console.log(state);
+  };
+
+  const getState = () => {
+    if (state === null) throw new Error('state not initialized');
+    return state;
+  };
+
+  const [setTimer, clearTimers] = createTimer();
+
+  const isSelected = (card: Card) => {
+    const { selectedCards } = getState();
+    return (
+      (selectedCards.first !== null && card.id === selectedCards.first.id) ||
+      (selectedCards.second !== null && card.id === selectedCards.second.id)
+    );
+  };
+
+  const setCards = (cards: Card[]) => {
+    setState((state) => ({ ...state, cards }));
+  };
+
+  const setScene = (scene: Scene) => {
+    setState((state) => ({ ...state, scene }));
+  };
+
+  const setSelectedCards = (selectedCards: SelectedCards) => {
+    setState((state) => ({ ...state, selectedCards }));
+  };
+
+  const disableControls = () => {
+    setState((state) => ({ ...state, controlsEnabled: false }));
+  };
+
+  const enableControls = () => {
+    setState((state) => ({ ...state, controlsEnabled: true }));
+  };
 
   const flipCard = (cardToFlip: Card) => {
-    setCards((cards) => cards.map((card) => (card.id === cardToFlip.id ? { ...card, flipped: !card.flipped } : card)));
+    console.log('flipp card');
+    setCards(getState().cards.map((card) => (card.id === cardToFlip.id ? { ...card, flipped: !card.flipped } : card)));
+  };
+
+  const initGame = (theme: Theme) => {
+    console.log('init game');
+    // init/replace game state with new data
+    const cards = createCardDeck(theme.cardsData);
+
+    const initState = {
+      theme,
+      cards,
+      selectedCards: { first: null, second: null },
+      scene: Scene.init,
+      controlsEnabled: false,
+    };
+
+    initializeState(initState);
+  };
+
+  const handleNewGameClick = () => {
+    console.log('new game click');
+    transitions[Transition.restart]();
   };
 
   const selectCard = (card: Card) => {
-    if (controlsDisabled || isSelected(card) || card.flipped) return;
+    const state = getState();
+    if (!state.controlsEnabled || isSelected(card) || card.flipped) return;
 
-    const { first, second } = selectedCards;
+    const { first, second } = state.selectedCards;
     // select first card
     if (first === null && !card.flipped) {
       transitions[Transition.firstCardSelected](card);
@@ -146,24 +223,25 @@ const MemoryGame = ({ theme }: MemoryGameProps) => {
   const transitions = {
     begin: () => {
       setScene(Scene.begin);
-      setControlsDisabled(true);
-      setCards(createCardDeck(cardsData));
-      setSelectedCards({ first: null, second: null });
+      disableControls();
+      updateUI();
       setTimer(() => {
         transitions[Transition.game]();
       }, 1000);
     },
     game: () => {
       setScene(Scene.game);
-      setControlsDisabled(false);
+      enableControls();
+      updateUI();
     },
     firstCardSelected: async (card: Card) => {
-      setControlsDisabled(true);
+      disableControls();
       setSelectedCards({ first: card, second: null });
       setScene(Scene.firstCardSelected);
       // play css animations and sounds
       flipCard(card); // 0.3s
-      await playAudio(audio.cardFlipSound);
+      updateUI();
+      await playAudio(getState().theme.audio.cardFlipSound);
       playCardPhrase(card);
 
       setTimer(() => {
@@ -171,12 +249,13 @@ const MemoryGame = ({ theme }: MemoryGameProps) => {
       }, 600); // reduced for fastrer UX
     },
     secondCardSelected: async ({ first, second }: { first: Card; second: Card }) => {
-      setControlsDisabled(true);
+      disableControls();
       setSelectedCards({ first, second });
       setScene(Scene.secondCardSelected);
       // play css animations and sounds
       flipCard(second); // 0.3s
-      await playAudio(audio.cardFlipSound);
+      updateUI();
+      await playAudio(getState().theme.audio.cardFlipSound);
       await playCardPhrase(second);
 
       const cardsMatch = first.image === second.image;
@@ -189,13 +268,15 @@ const MemoryGame = ({ theme }: MemoryGameProps) => {
     },
     cardsMatch: async () => {
       setScene(Scene.cardsMatch);
+      updateUI();
       delay(100);
-      await playAudio(audio.cardsMatchSound);
+      await playAudio(getState().theme.audio.cardsMatchSound);
       setScene(Scene.cardsMatchReward);
+      updateUI();
       Math.random() > 0.5 && (await playPhraseRandomLang(getRandomElement(phrases.good)));
       // reset selected cards
       // check win
-      if (cards.every((card) => card.flipped)) {
+      if (getState().cards.every((card) => card.flipped)) {
         transitions[Transition.win]();
       } else {
         setSelectedCards({ first: null, second: null });
@@ -204,46 +285,113 @@ const MemoryGame = ({ theme }: MemoryGameProps) => {
     },
     cardsDontMatch: async ({ first, second }: { first: Card; second: Card }) => {
       setScene(Scene.cardsDontMatch);
+      updateUI();
       await delay(1000);
       Math.random() > 0.8 && (await playPhraseRandomLang(getRandomElement(phrases.wrong)));
       flipCard(first);
       flipCard(second);
-      await playAudio(audio.cardFlipSound);
+      updateUI();
+      await playAudio(getState().theme.audio.cardFlipSound);
       setSelectedCards({ first: null, second: null });
       transitions[Transition.game]();
     },
     win: async () => {
       setScene(Scene.win);
+      updateUI();
       await playPhraseRandomLang(getRandomElement(phrases.win));
       await delay(200);
       setScene(Scene.winReward);
-      playAudio(audio.winMusic);
+      updateUI();
+      playAudio(getState().theme.audio.winMusic);
     },
     restart: async () => {
-      setControlsDisabled(true);
+      disableControls();
       playPhraseRandomLang(getRandomElement(phrases.newGame));
       setScene(Scene.goNewGame);
+      updateUI();
       clearTimers();
       setTimer(() => {
+        initGame(getState().theme);
         transitions[Transition.begin]();
       }, 500);
     },
   };
 
-  useEffect(() => {
-    console.log('useEffect theme fired');
-    transitions[Transition.begin]();
-    // clean up
-    return () => {
-      clearTimers();
-    };
-  }, [theme,clearTimers]);
+  interface Controls {
+    selectCard: (card: Card) => void;
+    isSelected: (card: Card) => boolean;
+    handleNewGameClick: () => void;
+  }
+
+  const controls = { selectCard, isSelected, handleNewGameClick };
+
+  interface Game {
+    state: State;
+    controls: Controls;
+  }
+
+  const getGame = () => (state === null ? null : { state, controls });
+
+  let uiCallback: null | (() => void) = null;
+  const connectUI = (func: () => void) => {
+    console.log('connect ui');
+    uiCallback = func;
+  };
+  const disconnectUI = () => (uiCallback = null);
+  const updateUI = () => {
+    if (uiCallback !== null) {
+      console.log('update ui');
+      uiCallback();
+    }
+  };
+
+  const useGame = (theme: Theme) => {
+    const [game, setGame] = useState<Game | null>(null);
+    const { playCardPhrase, playPhraseRandomLang } = usePlayPhrase();
+
+    useEffect(() => {
+      connectUI(() => setGame(getGame()));
+      return () => {
+        // disconnect ui
+        disconnectUI();
+      };
+    }, []);
+
+    useEffect(() => {
+      setPlayPhrase({ playCardPhrase, playPhraseRandomLang });
+    }, [playCardPhrase, playPhraseRandomLang]);
+
+    useEffect(() => {
+      // init with new data
+      initGame(theme);
+      transitions[Transition.begin]();
+      // start new game maybe
+    }, [theme]);
+
+    return game;
+  };
+
+  return useGame;
+};
+
+const useGame = createUseGame();
+
+const MemoryGame = ({ theme }: MemoryGameProps) => {
+  const { t } = useTranslation();
+
+  const game = useGame(theme);
+  if (game === null) return null;
+
+  const { scene, cards } = game.state;
+  const { image, styles } = game.state.theme;
+  const { isSelected, selectCard, handleNewGameClick } = game.controls;
 
   return (
     <div className={styles.app}>
-      <Button className={styles.newGameButton} text={t('utils.new_game')} onClick={transitions[Transition.restart]} />
+      <Button className={styles.newGameButton} text={t('utils.new_game')} onClick={handleNewGameClick} />
       <div className={styles.board}>
         {scene !== Scene.init &&
+          cards !== null &&
           cards.map((card) => (
             <Card
               key={card.id}
